+++ examples/data-sources/meraki_appliance_traffic_shaping_custom_performance_classes/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_traffic_shaping_custom_performance_classes" "example" {
+  network_id = "L_123456"
+}