+++ examples/resources/meraki_appliance_traffic_shaping_custom_performance_class/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_traffic_shaping_custom_performance_class.example "<network_id>,<id>"